refactor(cli-structure): move docs commands into docs package namespace
