+++ tests/unit/test_cli_mutation_boundary.py
@@ -6,7 +6,7 @@ from click.testing import CliRunner
 
 from local_rag_backend.app.contracts.results import MutationSummary
 from local_rag_backend.cli import cli
-from local_rag_backend.cli_commands import docs_ingest, docs_mutate
+from local_rag_backend.cli_commands.docs import docs_ingest, docs_mutate
 from local_rag_backend.settings import settings
